moved MockMediaManager to BaseTest
